Changes and improvements to energy usage graph
Graph can now display total energy usage for years and months when zoomed out using buttons. Fixed a bug were the graph would crash when zooming to far in or zooming in on the edges. Panning the graph when zoomed can be performed by holding down the shift key and dragging the graph.