Refactor frontend (#272)
* Updated styling of the comparison modal

* Started refactoring frontend

* Updated CSS

* Fixed layout and styling of comparisons

* Added percentage to the pie chart

* Moved some API calls to a shared file.

Moved the API calls for energy usage, energy saved and energy carriers to the file services/energy.tsx.

* Moved more API calls to own file

API calls for metrics and energy carriers for specific category has been moved to service files.

* Added comments to service files

* Fixed CSS

* Fixed color

* Added comment to random upgrades

* Started refactoring frontend

* Moved some API calls to a shared file.

Moved the API calls for energy usage, energy saved and energy carriers to the file services/energy.tsx.

* Moved more API calls to own file

API calls for metrics and energy carriers for specific category has been moved to service files.

* Added comments to service files

* Moved some API calls to a shared file.

Moved the API calls for energy usage, energy saved and energy carriers to the file services/energy.tsx.

* Moved more API calls to own file

API calls for metrics and energy carriers for specific category has been moved to service files.

* Moved all API calls to service files

* Added variable to get the base URI from the environment

* Fixed API_URI and added better typing to objects

* Moved energy usage fetching out of the component

Moved fetching the energy usage for the line graph out of the component. The data is sent as a prop to the component. This allows us to use reuse the graph in the comparison modal, instead of having two files.

* Moved energy usage fetching out of the component

Moved fetching the energy usage for the line graph out of the component. The data is sent as a prop to the component. This allows us to use reuse the graph in the comparison modal, instead of having two files.

* Fixed crashes when frontend is unable to find data

* Add cors and change backend port to 4000

* Added prefix to tooltip for energy usage

Co-authored-by: Thomas H <itmethomas@gmail.com>